@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -39,7 +39,7 @@ dependencies:
   json_serializable: ^4.1.1
   dio: ^4.0.0
   quiver: ^3.0.1
-  linked_scroll_controller: ^0.1.2
+  linked_scroll_controller: ^0.2.0-nullsafety.0
   flutter_svg: ^0.22.0
   cached_network_image: ^3.0.0
   image_picker: ^0.7.4
